auth basic

+++ model/request/basic.go
@@ -1,7 +1,7 @@
 package request
 
 type Basic struct {
-	Timestamp uint64 `json:"timestamp" binding:"required"`
+	Timestamp int `json:"timestamp" binding:"required"`
 	Token string `json:"token" binding:"required"`
 	Data map[string]interface{} `json:"data" binding:"required"`
 }
